pyproject.toml
[tool.poetry]
name = "observers-backend-django"
version = "0.1.0"
description = "Backend API for observers.gipss.tech written with Django and DRF"
authors = ["Gipss <gissa.dmtr@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "^3.13.1"
environs = {extras = ["django"], version = "^9.5.0"}
djangorestframework-simplejwt = "^5.2.0"
django-debug-toolbar = "^3.5.0"
drf-yasg = {extras = ["validation"], version = "^1.21.3"}
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.16.1"
autoflake = "^1.4"
yesqa = "^1.4.0"
pre-commit = "^2.20.0"
black = {extras = ["d"], version = "^22.6.0"}
mypy = "0.961"
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^1.7.0"}
django-stubs = {extras = ["compatible-mypy"], version = "^1.12.0"}


[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = [".",]

[tool.mypy]
pretty = true
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = false
strict_optional = true
strict_equality = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true
warn_no_return = true
exclude = ["manage.py", "migrations/"]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
